Fix major compilation errors: VastcoreLogger, ProBuilder API, TextMeshPro, IEnumerator types, Font conversion, DebugUI references
- Add Vastcore.Utils reference to Terrain assembly
- Add using Vastcore.Utils to affected files
- Replace Subdivide() calls with warning messages (ProBuilder API changes)
- Fix TextAlignmentOptions constants (MiddleLeft -> MidlineLeft, etc.)
- Add explicit System.Collections.IEnumerator type specifications
- Convert Font fields to TMP_FontAsset for TextMeshPro compatibility
- Temporarily disable DebugUI references with TODO comments

Resolved approximately 89 compilation errors out of 539 total.
