+++ Assets/Deform/EditorResources/DeformSettings.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:8c3a364cb0be1902a03c8cfd49ad29270f5078817c8e7aeed314c3119bdc38d5
-size 705
+oid sha256:07d5c850ab4208a20b67c0063cb4ed06058d6c1a7fac038cab0e5717a68120a6
+size 757
