+++ Assets/Scripts/Generation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3513256212908874898510348d2189e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
