+++ Assets/Scripts/Generation/PrimitiveTerrainObject.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 97a76a309c484744fb9d5002020fa51a
