+++ Assets/Scripts/Generation/Vastcore.Generation.asmdef.meta
