--- Assets/Scripts/Terrain/Map/PrimitiveTerrainGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 97367123ec6374b4bbab39e27af52189
