+++ Assets/Scripts/Terrain/Map/RuntimeTerrainManager.cs
@@ -1,7 +1,8 @@
 using UnityEngine;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
+using Vastcore.Player;
+using Vastcore.Utils;
 using Vastcore.Generation;
 using Vastcore.Core;
