+++ Packages/manifest.json
@@ -2,10 +2,12 @@
   "dependencies": {
     "com.beans.deform": "https://github.com/keenanwoodall/Deform.git",
     "com.justinpbarnett.unity-mcp": "https://github.com/justinpbarnett/unity-mcp.git?path=/UnityMcpBridge",
+    "com.unity.ai.assistant": "1.0.0-pre.12",
+    "com.unity.ai.generators": "1.0.0-pre.20",
+    "com.unity.ai.inference": "2.3.0",
     "com.unity.burst": "1.8.24",
     "com.unity.collab-proxy": "2.8.2",
     "com.unity.feature.mobile": "1.0.0",
-    "com.unity.ide.rider": "3.0.36",
     "com.unity.ide.visualstudio": "2.0.23",
     "com.unity.inputsystem": "1.14.2",
     "com.unity.multiplayer.center": "1.0.0",
